+++ packages/visual-stack-redux/src/components/SlidingPanel.js
@@ -11,6 +11,8 @@ import {
 } from '@cjdev/visual-stack/lib/components/SlidingPanel';
 import {
   toggleSlidingPanel,
+  expandFilterDropdown,
+  hideFilterDropdown,
   toggleFilterDropdown,
   setSlidingPanelActiveState,
 } from '../actions';
@@ -83,7 +85,9 @@ export class InternalSlidingPanelDropdown extends Component {
   constructor(props) {
     super(props);
     if (this.props.initialActive) {
-      this.props.onClick();
+      this.props.expandFilterDropdown();
+    } else {
+      this.props.hideFilterDropdown();
     }
   }
   render() {
@@ -116,6 +120,12 @@ export const SlidingPanelDropdown = connect(
     ),
   }),
   (dispatch, { id }) => ({
+    expandFilterDropdown() {
+      dispatch(expandFilterDropdown(id));
+    },
+    hideFilterDropdown() {
+      dispatch(hideFilterDropdown(id));
+    },
     onClick() {
       dispatch(toggleFilterDropdown(id));
     },