+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Upcoming
+# 4.9.2
 ## Bug Fixes
 -Fix dropdown filter's incorrect state when remounting with the same id